CurrentLoop

Coils.CurrentLoop — Type

CurrentLoop(; current::Unitful.Current, radius::Unitful.Length, height::Unitful.Length = 0u"m")

A circular current CurrentLoop.

Keywords

current::Unitful.Current: The current running through the loop.

radius::Unitful.Length: The radius of the current loop.

height::Unitful.Length: The height of the current loop, zero by default.

Coils.mfd — Method

mfd(c::CurrentLoop, ρ, z)

Computes the magnetic flux density according to analytical solution to the Biot-Savart law, see [1] and [2].

[1]: Simpson, James C. et al. “Simple Analytic Expressions for the Magnetic Field of a Circular Current Loop.” (2001).

[2]: Jang, Taehun, et al. "Off-axis magnetic fields of a circular loop and a solenoid for the electromagnetic induction of a magnetic pendulum."

Arguments

c::CurrentLoop: The CurrentLoop.

ρ::Unitful.Length: The radial coordinate.

z::Unitful.Length: The axial coordinate.

Returns

Vector{Unitful.MagneticFluxDensity}: The radial and axial magnetic flux density components.

Coils.mfdz — Method

mfdz(c::CurrentLoop[, z])

Computes the magnetic flux density along the z-axis according to [1].

[1]: https://de.wikipedia.org/wiki/Leiterschleife

Arguments

c::CurrentLoop: The CurrentLoop.

z::Unitful.Length: The axial coordinate by default equal to the height of current loop c.

Returns

Vector{Unitful.MagneticFluxDensity}: The radial and axial magnetic flux density components.