Helical

Coils.Helical — Type

Helical(; current, inner_radius, outer_radius, length, height, turns)

A helical coil with turns in radial and axial directions.

Fields

current::Unitful.Current: The current in the coil.

inner_radius::Unitful.Length: The inner radius of the coil.

outer_radius::Unitful.Length: The outer radius of the coil.

length::Unitful.Length: The (axial) length of the coil.

height::Unitful.Length: The height or axial offset of the coil.

radial_turns::Unsigned: The number of turns in the radial direction.

axial_turns::Unsigned: The number of turns in the axial direction.

Coils.Pancake — Function

Pancake(; current, inner_radius, outer_radius, turns, height)

A pancake coil which has only turns in the radial direction.

Keywords

current::Unitful.Current: The current in the coil.

inner_radius::Unitful.Length: The inner radius of the coil.

outer_radius::Unitful.Length: The outer radius of the coil.

turns::Unsigned: The number of turns in the radial direction.

height::Unitful.Length: The height or axial offset of the coil.

Return

Helical: A helical coil with only radial turns.

Coils.Solenoid — Function

Solenoid(; current, radius, length, turns, height)

A solenoid coil which has only turns in the axial direction.

Keywords

current::Unitful.Current: The current in the coil.

radius::Unitful.Length: The radius of the coil.

length::Unitful.Length: The (axial) length of the coil.

turns::Unsigned: The number of turns in the axial direction.

height::Unitful.Length: The height or axial offset of the coil.

Return

Helical: A helical coil with only axial turns.