Helmholtz

Coils.Helmholtz — Type

Helmholtz(; coil::Helical, separation::Unitful.Length = (coil.outer_radius + coil.inner_radius) / 2)
AntiHelmholtz(; coil::Helical, separation::Unitful.Length = √3 * (coil.outer_radius + coil.inner_radius) / 2

A(n) (anti-)Helmholtz coil configuration of Helical coils.

Keywords

coil::Helical: The Helical coil configuration.

separation::Unitful.Length: The separation between the coils by default uses the ideal separation.

Missing docstring.

Missing docstring for Coils.AntiHelmholtz. Check Documenter's build log for details.

Warning: The values for the inductances have not been cross-checked!

Coils.self_inductance — Method

self_inductance(c::Helmholtz)

Computes an approximate self inductance of each of the coils in (anti-)Helmholtz configuration.

Dispatches the computation to total_inductance of the coil configuration.

Arguments

c::Helmholtz: The (anti-)Helmholtz coil.

Returns

Unitful.Inductance: The self inductance of each of the coils in (anti-)Helmholtz configuration.

Coils.mutual_inductance — Method

mutual_inductance(c::Helmholtz)

Computes an approximate mutual inductance between the coils of the (anti-)Helmholtz pair according to Ref. [1]

[1]: https://de.wikipedia.org/wiki/Helmholtz-Spule#Induktivit%C3%A4t

Arguments

c::Helmholtz: The (anti-)Helmholtz coil.

Returns

Unitful.Inductance: The mutual inductance between the coils of the (anti-)Helmholtz pair.

Missing docstring.

Missing docstring for Coils.total_inductance(c::Helmholtz). Check Documenter's build log for details.

Coils.mfdz — Method

mfdz(c::Helmholtz)

Computes the magnetic flux density at the center of an infinite-length (anti-)Helmholtz coil according to [1].

[1]: https://en.wikipedia.org/wiki/Helmholtz_coil

We use the formula of an infinite-length solenoid where we use the product of radial and axial turns as the number of turns.

Arguments

c::Helmholtz: The (anti-)Helmholtz coil.

Returns

Vector{Unitful.MagneticFluxDensity}: The radial and axial magnetic flux density components.